build: update gradlew scripts
Updated the `gradlew` script to remove unused code related to `APP_NAME` and `DEFAULT_JVM_OPTS`. Additionally, added a check for the availability of `xargs`.

In the `gradlew.bat` script, modified the error level check to use the `equ` operator instead of `==`. Also, set the `EXIT_CODE` variable and modified the exit code for failed execution in order to return the correct return code.

These changes improve the efficiency and maintainability of the scripts.