ci: refactor the build workflow to use Gradle instead of Java
The `.github/workflows/build.yml` file has been modified to replace the setup of Java 11 environment with the setup of Gradle. The steps for setting up Java have been removed, and a new step for setting up Gradle has been added. This change allows the build workflow to use Gradle instead of Java for the subsequent steps.