ci: add plugin building and artifact creation steps
Added new steps to the workflow to build the plugin and create an artifact from it. The plugin archive content is unzipped for further processing, and the already-built plugin is stored as an artifact for downloading.

Also, added a new verification step to run plugin structure verification and use IntelliJ Plugin Verifier. Unnecessary files are removed, the repository is checked out, Gradle is set up, and Plugin Verifier IDEs are cached. Finally, the plugin verification task is run, and the result is collected.

Additionally, implemented a new step to run Qodana inspections on the code. This step is skipped if the actor is 'dependabot[bot]'.