build: update Gradle version and improve Java check in gradlew
This commit includes the upgrade of Gradle from 8.1.1 to 8.2.1, which includes various performance improvements and bug fixes. Additionally, it refactors the Java check in the gradlew script to use a more reliable command syntax.

A new property `validateDistributionUrl` is also added to the gradle-wrapper.properties file which helps to validate the downloaded Gradle distribution.

Lastly, a newline character is added at the end of settings.gradle.kts for better file formatting.